Compile everything with -Werror, FE-1757.

+++ src/Widgets/AE/ColorAEWidget.cpp
@@ -40,7 +40,9 @@ void ColorPickerWidget::mousePressEvent ( QMouseEvent * event )
   QColorDialog dialog( this );
   dialog.setOptions(QColorDialog::ShowAlphaChannel);
   dialog.setModal(true);
-  dialog.setCurrentColor(QColor(m_colorR * 255.0f, m_colorG * 255.0f, m_colorB * 255.0f, m_colorA * 255.0f));
+  dialog.setCurrentColor(
+    QColor( int( m_colorR * 255 ), int( m_colorG * 255 ),
+            int( m_colorB * 255 ), int( m_colorA * 255 ) ) );
   connect(&dialog, SIGNAL(rejected()), this, SLOT(colorDialogRejected()));
   connect(&dialog, SIGNAL(currentColorChanged(const QColor &)), this, SLOT(colorDialogChanged(const QColor &)));
 
@@ -61,7 +63,10 @@ void ColorPickerWidget::paintEvent ( QPaintEvent * event )
 {
   QPainter painter;
   painter.begin(this);
-  painter.fillRect(event->rect(), QBrush(QColor(m_colorR * 255.0f, m_colorG * 255.0f, m_colorB * 255.0f)));
+  painter.fillRect(
+    event->rect(),
+    QBrush( QColor( int( m_colorR * 255 ), int( m_colorG * 255 ),
+                    int( m_colorB * 255 ) ) ) );
   painter.end();  
 }
